refactor(session): pass cookie session id to service
