+++ packages/server/src/aggregates/session/domain/session.abstract.service.ts
@@ -1,7 +1,6 @@
-import { Request } from 'express';
 import { SessionEntity } from '@rateme/core/domain/entities/session.entity';
 
 export abstract class SessionAbstractService {
-  abstract getSession(request: Request): Promise<SessionEntity>;
+  abstract getSession(sessionId: string): Promise<SessionEntity>;
 }
