+++ packages/server/src/aggregates/session/infrastructure/session.service.ts
@@ -1,8 +1,6 @@
 import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
-import { Request } from 'express';
 
 import { SessionStatus } from '@rateme/core/domain/entities/session.entity';
-import { CookieService } from '@/core/modules/cookie';
 
 import { SessionAbstractService } from '../domain';
 import { SessionUnitOfWork } from './session.unit-of-work';
@@ -12,18 +10,11 @@ export class SessionService extends SessionAbstractService {
   constructor(
     @Inject(SessionUnitOfWork)
     private readonly sessionUnitOfWork: SessionUnitOfWork,
-    @Inject(CookieService)
-    private readonly cookieService: CookieService,
   ) {
     super();
   }
 
-  async getSession(request: Request) {
-    const sessionId = this.cookieService.getSessionId(request);
-
-    if (!sessionId) {
-      throw new UnauthorizedException();
-    }
+  async getSession(sessionId: string) {
 
     return this.sessionUnitOfWork.start(async ({ sessionRepository }) => {
       const session = await sessionRepository.findById(sessionId);
